requirements.txt
mysql-python==1.2.5
flask==0.12.2
flask_mail==0.9.1
flask_sqlalchemy==2.3.2
flask_login==0.4.0
flask_script==2.0.6
flask_migrate==2.1.1
enum34==1.1.6
flask_wtf==0.14.2
requests==2.18.4
flask_restful==0.3.6
celery==4.1.0
redis==2.10.6
gevent==1.2.2
celery==4.1.0
colorlog==3.1.0
